Tests comments and move http server handlers from cas_server.utils to cas_server.tests.utils
