+++ cas_server/utils.py
@@ -23,10 +23,8 @@ import hashlib
 import crypt
 import base64
 import six
-import cgi
-from threading import Thread
+
 from importlib import import_module
-from six.moves import BaseHTTPServer
 from six.moves.urllib.parse import urlparse, urlunparse, parse_qsl, urlencode
 
 
@@ -149,70 +147,6 @@ def gen_pgtiou():
 def gen_saml_id():
     """Generate an saml id"""
     return _gen_ticket('_')
-
-
-class HttpParamsHandler(BaseHTTPServer.BaseHTTPRequestHandler):
-    """
-        A simple http server that return 200 on GET or POST
-        and store GET or POST parameters. Used in unit tests
-    """
-
-    def do_GET(self):
-        """Called on a GET request on the BaseHTTPServer"""
-        self.send_response(200)
-        self.send_header(b"Content-type", "text/plain")
-        self.end_headers()
-        self.wfile.write(b"ok")
-        url = urlparse(self.path)
-        params = dict(parse_qsl(url.query))
-        self.server.PARAMS = params
-
-    def do_POST(self):
-        """Called on a POST request on the BaseHTTPServer"""
-        ctype, pdict = cgi.parse_header(self.headers.get('content-type'))
-        if ctype == 'multipart/form-data':
-            postvars = cgi.parse_multipart(self.rfile, pdict)
-        elif ctype == 'application/x-www-form-urlencoded':
-            length = int(self.headers.get('content-length'))
-            postvars = cgi.parse_qs(self.rfile.read(length), keep_blank_values=1)
-        else:
-            postvars = {}
-        self.server.PARAMS = postvars
-
-    def log_message(self, *args):
-        """silent any log message"""
-        return
-
-    @classmethod
-    def run(cls):
-        """Run a BaseHTTPServer using this class as handler"""
-        server_class = BaseHTTPServer.HTTPServer
-        httpd = server_class(("127.0.0.1", 0), cls)
-        (host, port) = httpd.socket.getsockname()
-
-        def lauch():
-            """routine to lauch in a background thread"""
-            httpd.handle_request()
-            httpd.server_close()
-
-        httpd_thread = Thread(target=lauch)
-        httpd_thread.daemon = True
-        httpd_thread.start()
-        return (httpd, host, port)
-
-
-class Http404Handler(HttpParamsHandler):
-    """A simple http server that always return 404 not found. Used in unit tests"""
-    def do_GET(self):
-        """Called on a GET request on the BaseHTTPServer"""
-        self.send_response(404)
-        self.send_header(b"Content-type", "text/plain")
-        self.end_headers()
-        self.wfile.write(b"error 404 not found")
-
-    def do_POST(self):
-        """Called on a POST request on the BaseHTTPServer"""
-        return self.do_GET()
 
 
 class LdapHashUserPassword(object):
